code reorganization: extracted ModelGenerator into separate files
* moved shared functions between wrap and ModelGenerator into utils.cpp
* moved PRTContext singleton into PRTContext.cpp and added mgmt functions
