+++ src/client/PRTContext.cpp
@@ -2,6 +2,26 @@
 
 #include <array>
 #include <string>
+#include <memory>
+#include <mutex>
+
+namespace {
+
+std::shared_ptr<PRTContext> prtCtx;
+std::once_flag prtInitFlag;
+
+} // namespace
+
+std::shared_ptr<PRTContext> PRTContext::get() {
+	std::call_once(prtInitFlag, [](){
+		prtCtx = std::make_shared<PRTContext>(prt::LOG_WARNING);
+	});
+	return prtCtx;
+}
+
+void PRTContext::shutdown() {
+	prtCtx.reset();
+}
 
 PRTContext::PRTContext(prt::LogLevel minimalLogLevel) {
 	prt::addLogHandler(&mLogHandler);
@@ -23,8 +43,4 @@ PRTContext::~PRTContext() {
 
 	// remove loggers
 	prt::removeLogHandler(&mLogHandler);
-}
-
-PRTContext::operator bool() const {
-	return (bool)mPRTHandle;
 }
