+++ src/client/PRTContext.h
@@ -9,10 +9,11 @@
  * Helper struct to manage PRT lifetime (e.g. the prt::init() call)
  */
 struct PRTContext {
+	static std::shared_ptr<PRTContext> get();
+	static void shutdown();
+
 	PRTContext(prt::LogLevel minimalLogLevel);
 	~PRTContext();
-
-	explicit operator bool() const;
 
 	PythonLogHandler mLogHandler;
 	pcu::ObjectPtr mPRTHandle;
