+++ src/client/utils.cpp
@@ -60,6 +60,47 @@ void tokenize(const std::basic_string<C>& str, std::vector<std::basic_string<C>>
 namespace py = pybind11;
 
 namespace pcu {
+
+bool getResolveMap(const std::filesystem::path& rulePackagePath, ResolveMapPtr* resolveMap) {
+	if (std::filesystem::exists(rulePackagePath)) {
+		LOG_INF << "using rule package " << rulePackagePath << std::endl;
+
+		const std::string u8rpkURI = toFileURI(rulePackagePath.string());
+		prt::Status status = prt::STATUS_UNSPECIFIED_ERROR;
+		try {
+			resolveMap->reset(prt::createResolveMap(toUTF16FromUTF8(u8rpkURI).c_str(), nullptr, &status));
+		}
+		catch (const std::exception& e) {
+			pybind11::print("CAUGHT EXCEPTION:", e.what());
+			return false;
+		}
+
+		if (resolveMap && (status == prt::STATUS_OK)) {
+			LOG_DBG << "resolve map = " << objectToXML(resolveMap->get()) << std::endl;
+		}
+		else {
+			LOG_ERR << "getting resolve map from '" << rulePackagePath << "' failed, aborting.";
+			return false;
+		}
+	}
+
+	return true;
+}
+
+std::wstring getRuleFileEntry(const prt::ResolveMap* resolveMap) {
+	const std::wstring sCGB(L".cgb");
+
+	size_t nKeys;
+	wchar_t const* const* keys = resolveMap->getKeys(&nKeys);
+	for (size_t k = 0; k < nKeys; k++) {
+		const std::wstring key(keys[k]);
+		if (std::equal(sCGB.rbegin(), sCGB.rend(), key.rbegin()))
+			return key;
+	}
+
+	return {};
+}
+
 /**
  * Helper function to convert a Python dictionary of "<key>:<value>" into a
  * prt::AttributeMap
