+++ src/client/wrap.h
@@ -62,44 +62,3 @@ inline void copyToCStr(const std::string& str, char* cstr, size_t& cstrSize) {
 	}
 	cstrSize = str.length() + 1; // returns the actually needed size including terminating null
 }
-
-
-namespace {
-
-class ModelGenerator {
-public:
-	ModelGenerator(const std::vector<InitialShape>& myGeo);
-	~ModelGenerator() {}
-
-	std::vector<GeneratedModel> generateModel(const std::vector<py::dict>& shapeAttributes,
-	                                          const std::filesystem::path& rulePackagePath,
-	                                          const std::wstring& geometryEncoderName,
-	                                          const py::dict& geometryEcoderOptions);
-	std::vector<GeneratedModel> generateAnotherModel(const std::vector<py::dict>& shapeAttributes);
-
-private:
-	pcu::ResolveMapPtr mResolveMap;
-	pcu::CachePtr mCache;
-
-	pcu::AttributeMapBuilderPtr mEncoderBuilder;
-	std::vector<pcu::AttributeMapPtr> mEncodersOptionsPtr;
-	std::vector<std::wstring> mEncodersNames;
-	std::vector<pcu::InitialShapeBuilderPtr> mInitialShapesBuilders;
-
-	std::wstring mRuleFile = L"";
-	std::wstring mStartRule = L"";
-	int32_t mSeed = 0;
-	std::wstring mShapeName = L"InitialShape";
-
-	bool mValid = true;
-
-	void setAndCreateInitialShape(const std::vector<py::dict>& shapeAttr,
-	                              std::vector<const prt::InitialShape*>& initShapes,
-	                              std::vector<pcu::InitialShapePtr>& initShapesPtrs,
-	                              std::vector<pcu::AttributeMapPtr>& convertShapeAttr);
-	void initializeEncoderData(const std::wstring& encName, const py::dict& encOpt);
-	void getRawEncoderDataPointers(std::vector<const wchar_t*>& allEnc,
-	                               std::vector<const prt::AttributeMap*>& allEncOpt);
-};
-
-} // namespace
